% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_extract.R
\name{etl_extract.etl_macleish}
\alias{etl_extract.etl_macleish}
\alias{etl_load.etl_macleish}
\alias{etl_transform.etl_macleish}
\title{Extract weather data}
\usage{
\method{etl_extract}{etl_macleish}(obj, ...)

\method{etl_transform}{etl_macleish}(obj, ...)

\method{etl_load}{etl_macleish}(obj, ...)
}
\description{
Extract weather data
}
\examples{

macleish <- etl("macleish")
str(macleish)

macleish \%>\%
  etl_extract() \%>\%
  etl_transform() \%>\%
  etl_load()
whately <- macleish \%>\%
  tbl("whately") 
  
whately \%>\%
  summarize(N = n(), avg_temp = mean(Temp_C_Avg))
}

