% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_mdata.R
\name{get_mdata}
\alias{get_mdata}
\title{get_mdata}
\usage{
get_mdata()
}
\description{
Retrieve data from the Macleish Field Station weather monitors
}
\examples{
macleish <- get_mdata()
if (require(lattice)) {
 xyplot(Temp_C_Avg ~ when, data = macleish, type = c("l", "smooth"))
}
}

