% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdata.R
\name{get_orchard}
\alias{get_orchard}
\title{get_orchard}
\usage{
get_orchard()
}
\description{
Retrieve data from the Macleish Field Station weather monitors
}
\examples{

orchard <- get_orchard()
if (require(lattice)) {
 xyplot(Temp_C_Avg ~ when, data = orchard, type = c("l", "smooth"))
}
}

