% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_mdata.R
\name{get_whately}
\alias{get_whately}
\title{get_whately}
\usage{
get_whately()
}
\description{
Retrieve data from the Macleish Field Station weather monitors
}
\examples{
whately <- get_whately()
if (require(lattice)) {
 xyplot(Temp_C_Avg ~ when, data = whately, type = c("l", "smooth"))
}
}

