% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_gis.R
\name{macleish_elevation}
\alias{macleish_elevation}
\title{Retrieve elevation layers from MassGIS}
\usage{
macleish_elevation()
}
\description{
Retrieve elevation layers from MassGIS
}
\details{
This function will download shapefiles from MassGIS, unzip them, 
transform the projection to EPSG:4326, compute their intersection with the
boundary of the MacLeish property, and return the resulting 
\code{\link[sp]{SpatialLines-class}} object.
}
\examples{

\dontrun{
# have to download the shapefiles...cound take a while...
elevation <- macleish_elevation()
if (require(sp)) {
  plot(elevation)
}
}
}
