% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maple_sap}
\alias{maple_sap}
\title{Maple sap collection}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 118 rows and 4 columns.}
\usage{
maple_sap
}
\description{
Maple sap collection
}
\keyword{datasets}
