% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_data02}
\alias{plot_data02}
\title{Plot data collection at MacLeish}
\format{
\describe{
 \item{when}{the date of collection}
 \item{sap}{how much sap was collected, in gallons}
 \item{Comments}{comments}
 \item{People}{who was there?}
}
}
\usage{
plot_data02
}
\description{
Plot data collection at MacLeish
}
\keyword{datasets}
