% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree_diameter2}
\alias{tree_diameter2}
\title{MacLeish Data Plot 2}
\format{
\describe{
 \item{Module}{Module number that represents one of the 10 subplot modules that are 110 m
 and 20 x 50 m. There are five 10 x 10 modules along central 50 m axis.}
 \item{Tag}{Tag numbers used to identify each tree.}
 \item{Species}{Tree species include Acer rubrum, Betula lenta, Fagus grandifolia, Pinus strobus, 
 Quercus rubra, and Tsuga canadensis.}
 \item{Year}{Data collected from 2009-2012 in the fall semesters.}
 \item{Notes}{Notes collected on the wellbeing of tree species, including notes on death or poor health.}
 \item{Diameter}{Tree diameter measured in centimeters and at 1.4 meters high above the ground.}
 \item{People}{Data obtained from Jesse Bellemare and Smith College students from BIO364-365 courses.}
}
}
\usage{
tree_diameter2
}
\description{
Data on change in tree diameter (in centimeters) 
for parasitic Hemlock Woolly Adelgid dominated areas on the Western side of MacLeish. 
Tree diameter was measured at 1.4 meters high above the ground.
}
\keyword{datasets}
