% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{whately}
\alias{whately}
\title{Weather data from Macleish Field Station}
\format{A data frame with several hundred thousand row and 9 variables:
\describe{
  \item{when}{timestamp, recorded every 10 minutes}
  \item{record}{?}
  \item{Temp_C_Avg}{average temperature, in Celsius, over these 10 minutes}
}}
\source{
These data are recorded at \url{http://www.smith.edu/ceeds/macleish.php}
}
\usage{
whately
}
\description{
data from Macleish from 2010 to present.
}
\details{
the description
}
\keyword{datasets}

