% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macleish_layers}
\alias{macleish_layers}
\title{MacLeish spatial data}
\format{A \code{list} of \code{\link[sp]{Spatial}} objects, each providing a different layer.

\describe{
 \item{buildings}{Buildings at MacLeish}
}}
\usage{
macleish_layers
}
\description{
Shapefiles from the MacLeish Field Station
}
\keyword{datasets}

