% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_data01}
\alias{plot_data01}
\title{Plot data collection at MacLeish}
\format{
\describe{
 \item{when}{Data collected from 2010-2015}
 \item{size}{20 x 50 m and divided into 10 subplot modules of 110 m} 
 \item{Comments}{comments}
 \item{People}{Jesse Bellemare}
}
}
\usage{
plot_data01
}
\description{
Plot data collection at MacLeish
}
\keyword{datasets}
