% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plot_data01}
\alias{plot_data01}
\title{MacLeish Data Plot 1}
\format{
\describe{
 \item{When}{Data collected from 2010-2015 in the fall semesters.}
 \item{Size}{20 x 50 m and divided into 10 subplot modules of 110 m. Five 10 x 10 modules along central 50 m axis.} 
 \item{Comments}{comments}
 \item{People}{Data obtained from Jesse Bellemare and Smith College students from BIO364-365 course.}
}
}
\usage{
plot_data01
}
\description{
Collected data on the trend of tree diameter in centimeters 
for Hemlock Woolly Adelgid dominated areas on the Western side of MacLeish. 
Tree diameter was measured at 1.4 meters high above the ground.
}
\keyword{datasets}
